rework to use sqlite instead of csv file
