+++ requirements.txt
@@ -1,0 +1,6 @@
+googlemaps==4.5.3
+folium==0.12.1
+pandas==1.3.3
+halo==0.0.31
+requests==2.26.0
+plotly==5.3.1
